Ability to activate Missile & Laser weapon
Added functionality to initialize the missile addon from a prefab
Added a pretty cool laser weapon that does damage over time.

+++ Assets/Scripts/Enemy/EnemyDeath.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class EnemyDeath : MonoBehaviour
 {
-    public int health = 1;
+    public float health = 1;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
         if (health <= 0)
